refactor(electron-example): clean up comments & types
