+++ apps/electron-example/src/main/store.ts
@@ -1,14 +1,15 @@
 import { createStore } from 'zustand/vanilla';
 import { getZubridgeMode } from '../utils/mode.js';
 
-import type { State } from '../types/state.js';
+import type { State } from '../types/index.js';
 
-// Create store with the appropriate initial state based on mode
+/**
+ * Creates a store with appropriate initial state for the current mode
+ */
 export const createModeStore = () => {
   const mode = getZubridgeMode();
   console.log(`[Store] Creating store for mode: ${mode}`);
 
-  // Common initial state for all modes
   const initialState = {
     counter: 0,
     window: { isOpen: false },
